V2 Spike 2: Empfangspfad Datei -> Sync -> OFDM -> FIC Qt-frei in libdabcore
- Dateiquellen (.uff mit eigenem XML-Header-Parser, rohe int8-.iq), Pacing
  in Echtzeit oder --fast, --duration in Dateizeit, --loop, file_progress/file_ended
- Frontend portiert aus Qt-DAB (sample-reader, timesyncer, correlator, freqsyncer,
  phasetable, estimator, freq-interleaver, ofdm-decoder, ofdm-handler, TII),
  FIC komplett (fic-handler, fib-decoder inkl. FIG 0/15 EWS und FIG 0/19,
  fib-config, ensemble), DSP (ringbuffer, FFT, Viterbi AVX2, protection, CRC,
  Tabellen), charsets neu ohne Qt; MSC als No-op-Senke
- DabCore: open_device{file}, Ereignisse synced/snr/fic_quality/ensemble_found/
  service_added/clock_time/ews_*/tii; dab-api und dab-cli um fast/ews_alive.sub_ch
- Race in ofdmHandler behoben: setRunning(true) jetzt in start() statt run(),
  sonst haengt stop() bei sofortigem Beenden
- ctest dabcored_replay am 60-s-Referenzfile (Ensemble, Dienste, EWS-Heartbeat);
  60-s-Datei ohne Pacing in 3,5 s; Warntag-Alarm: pre_trigger/trigger/sustain/end

+++ V2/core-cpp/libdabcore/include/dabcore/events.h
@@ -74,7 +74,8 @@ json audioDevices(const std::vector<std::string>& names, int current);
 
 json ewsPresent();
 json ewsAlert(EwsPhase phase, uint8_t subCh, uint8_t stage, uint16_t iid, const std::vector<std::string>& locations, bool isTest);
-json ewsAlive(uint8_t subCh);
+// subCh < 0: Heartbeat ohne aktiven Alarm (sub_ch = null)
+json ewsAlive(int subCh);
 json ewfAlarm(bool active, uint8_t subCh);
 json ewsSwitched(uint32_t toSid, int64_t fromSid /* <0 = keiner */);
